refactor(config): moved structs into seperate modules
